Exceptional.jl Documentation

Exceptional — Module

Exceptional signals and handles exceptional situations, including use of restarts. An exceptional situation occur when a program reaches a point where a planned operation cannot be done.

Main.Exceptional.block — Method

block makes it possible to do a non-local transfer of control. By using block, a named exit point is set. Calling the return_from function inside the block context make it possible to return values from those named exit points.

Main.Exceptional.return_from — Function

return_from works together with the block function. It's telling the block function which value to return, and from which context this value should be returned.

Main.Exceptional.available_restart — Method

available_restart takes a named restarts as input and returns true if this is a possible restart to do.

Main.Exceptional.invoke_restart — Method

invoke_restart finds the registered restart connected to the input argument name. It returns the corresponding restart function with corresponding arguments.

Main.Exceptional.restart_bind — Method

restart_bind gives the ability to go back to the place where an exceptional situation happened, from this place a given action called restart can be done to fix the occuring error.

Main.Exceptional.error — Method

error recives an exception as input. It check the registered handlers, if any of these handlers is the same type as the input exception, or a parent of the input exception, that handler function is called. If none of the handler functions returns a result the incoming error is thrown.

Main.Exceptional.handler_bind — Method

handler_bind makes it possible to be notified when an exceptional situation occur. This is done by wrapping a function who could give an exceptional situation within handlers.